requirements.txt
Flask==1.0.2
Flask-Cors==3.0.6
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gunicorn==20.0.4
ipython==7.10.1
ipython-genutils==0.2.0
ipywidgets==7.4.1
itsdangerous==0.24
Jinja2==2.10
path.py==11.1.0
pathlib2==2.3.2
patsy==0.5.0
psutil==5.4.7
psycopg2==2.8.4
ptyprocess==0.6.0
py==1.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0.5
pyflakes==2.0.0
Pygments==2.2.0
PyHamcrest==1.9.0
pylint==2.1.1
pylint-flask==0.6
pylint-plugin-utils==0.6
pyOpenSSL==18.0.0
pyparsing==2.2.0
PySocks==1.6.8
pytest==3.8.0
pytest-arraydiff==0.2
pytest-astropy==0.4.0
pytest-doctestplus==0.1.3
pytest-openfiles==0.3.0
pytest-remotedata==0.3.0
python-dateutil==2.7.3
python-dotenv==0.13.0
pyzmq==17.1.2
SQLAlchemy==1.2.11
Werkzeug==0.14.1
WTForms==2.3.1